requirements.txt
beautifulsoup4==4.9.1
fake-useragent==0.1.11
pandas==1.1.1
selenium==3.141.0
webdriver-manager==3.2.2
PyYAML==5.3.1